ZCS-806 Add flags to ldap sieve attributes
