+++ store/src/java-test/com/zimbra/cs/filter/EscapeSequencesTest.java
@@ -70,7 +70,6 @@ public final class EscapeSequencesTest {
         attrs.put(Provisioning.A_zimbraId, UUID.randomUUID().toString());
         Account account = prov.createAccount("test1@zimbra.com", "secret", attrs);
         Server server = Provisioning.getInstance().getServer(account);
-        server.setSieveFeatureVariablesEnabled(true);
 
         attrs = Maps.newHashMap();
         attrs.put(Provisioning.A_zimbraId, UUID.randomUUID().toString());
