+++ store/src/java/com/zimbra/cs/account/ZAttrConfig.java
@@ -2074,83 +2074,6 @@ public abstract class ZAttrConfig extends Entry {
     }
 
     /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the admin-defined sieve rules.
-     *
-     * @return zimbraAdminSieveFeatureVariablesEnabled, or true if unset
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2098)
-    public boolean isAdminSieveFeatureVariablesEnabled() {
-        return getBooleanAttr(Provisioning.A_zimbraAdminSieveFeatureVariablesEnabled, true, true);
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the admin-defined sieve rules.
-     *
-     * @param zimbraAdminSieveFeatureVariablesEnabled new value
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2098)
-    public void setAdminSieveFeatureVariablesEnabled(boolean zimbraAdminSieveFeatureVariablesEnabled) throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraAdminSieveFeatureVariablesEnabled, zimbraAdminSieveFeatureVariablesEnabled ? Provisioning.TRUE : Provisioning.FALSE);
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the admin-defined sieve rules.
-     *
-     * @param zimbraAdminSieveFeatureVariablesEnabled new value
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2098)
-    public Map<String,Object> setAdminSieveFeatureVariablesEnabled(boolean zimbraAdminSieveFeatureVariablesEnabled, Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraAdminSieveFeatureVariablesEnabled, zimbraAdminSieveFeatureVariablesEnabled ? Provisioning.TRUE : Provisioning.FALSE);
-        return attrs;
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the admin-defined sieve rules.
-     *
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2098)
-    public void unsetAdminSieveFeatureVariablesEnabled() throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraAdminSieveFeatureVariablesEnabled, "");
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the admin-defined sieve rules.
-     *
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2098)
-    public Map<String,Object> unsetAdminSieveFeatureVariablesEnabled(Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraAdminSieveFeatureVariablesEnabled, "");
-        return attrs;
-    }
-
-    /**
      * URL prefix for where the zimbraAdmin app resides on this server
      *
      * @return zimbraAdminURL, or "/zimbraAdmin" if unset
@@ -9043,7 +8966,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2106)
     public boolean isChatAllowUnencryptedPassword() {
-        return getBooleanAttr(Provisioning.A_zimbraChatAllowUnencryptedPassword, false);
+        return getBooleanAttr(Provisioning.A_zimbraChatAllowUnencryptedPassword, false, true);
     }
 
     /**
@@ -9116,7 +9039,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2104)
     public boolean isChatConversationAuditEnabled() {
-        return getBooleanAttr(Provisioning.A_zimbraChatConversationAuditEnabled, false);
+        return getBooleanAttr(Provisioning.A_zimbraChatConversationAuditEnabled, false, true);
     }
 
     /**
@@ -9192,7 +9115,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2102)
     public boolean isChatServiceEnabled() {
-        return getBooleanAttr(Provisioning.A_zimbraChatServiceEnabled, true);
+        return getBooleanAttr(Provisioning.A_zimbraChatServiceEnabled, true, true);
     }
 
     /**
@@ -9268,7 +9191,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2107)
     public int getChatXmppPort() {
-        return getIntAttr(Provisioning.A_zimbraChatXmppPort, 5222);
+        return getIntAttr(Provisioning.A_zimbraChatXmppPort, 5222, true);
     }
 
     /**
@@ -9280,7 +9203,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2107)
     public String getChatXmppPortAsString() {
-        return getAttr(Provisioning.A_zimbraChatXmppPort, "5222");
+        return getAttr(Provisioning.A_zimbraChatXmppPort, "5222", true);
     }
 
     /**
@@ -9387,7 +9310,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2108)
     public int getChatXmppSslPort() {
-        return getIntAttr(Provisioning.A_zimbraChatXmppSslPort, 5223);
+        return getIntAttr(Provisioning.A_zimbraChatXmppSslPort, 5223, true);
     }
 
     /**
@@ -9399,7 +9322,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2108)
     public String getChatXmppSslPortAsString() {
-        return getAttr(Provisioning.A_zimbraChatXmppSslPort, "5223");
+        return getAttr(Provisioning.A_zimbraChatXmppSslPort, "5223", true);
     }
 
     /**
@@ -9502,7 +9425,7 @@ public abstract class ZAttrConfig extends Entry {
      */
     @ZAttr(id=2105)
     public boolean isChatXmppSslPortEnabled() {
-        return getBooleanAttr(Provisioning.A_zimbraChatXmppSslPortEnabled, false);
+        return getBooleanAttr(Provisioning.A_zimbraChatXmppSslPortEnabled, false, true);
     }
 
     /**
@@ -25422,6 +25345,314 @@ public abstract class ZAttrConfig extends Entry {
     }
 
     /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied after the end user filter rule
+     *
+     * @return zimbraMailAdminOutgoingSieveScriptAfter, or null if unset
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2093)
+    public String getMailAdminOutgoingSieveScriptAfter() {
+        return getAttr(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, null, true);
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied after the end user filter rule
+     *
+     * @param zimbraMailAdminOutgoingSieveScriptAfter new value
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2093)
+    public void setMailAdminOutgoingSieveScriptAfter(String zimbraMailAdminOutgoingSieveScriptAfter) throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, zimbraMailAdminOutgoingSieveScriptAfter);
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied after the end user filter rule
+     *
+     * @param zimbraMailAdminOutgoingSieveScriptAfter new value
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2093)
+    public Map<String,Object> setMailAdminOutgoingSieveScriptAfter(String zimbraMailAdminOutgoingSieveScriptAfter, Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, zimbraMailAdminOutgoingSieveScriptAfter);
+        return attrs;
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied after the end user filter rule
+     *
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2093)
+    public void unsetMailAdminOutgoingSieveScriptAfter() throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, "");
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied after the end user filter rule
+     *
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2093)
+    public Map<String,Object> unsetMailAdminOutgoingSieveScriptAfter(Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptAfter, "");
+        return attrs;
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied before the end user filter rule
+     *
+     * @return zimbraMailAdminOutgoingSieveScriptBefore, or null if unset
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2092)
+    public String getMailAdminOutgoingSieveScriptBefore() {
+        return getAttr(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, null, true);
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied before the end user filter rule
+     *
+     * @param zimbraMailAdminOutgoingSieveScriptBefore new value
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2092)
+    public void setMailAdminOutgoingSieveScriptBefore(String zimbraMailAdminOutgoingSieveScriptBefore) throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, zimbraMailAdminOutgoingSieveScriptBefore);
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied before the end user filter rule
+     *
+     * @param zimbraMailAdminOutgoingSieveScriptBefore new value
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2092)
+    public Map<String,Object> setMailAdminOutgoingSieveScriptBefore(String zimbraMailAdminOutgoingSieveScriptBefore, Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, zimbraMailAdminOutgoingSieveScriptBefore);
+        return attrs;
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied before the end user filter rule
+     *
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2092)
+    public void unsetMailAdminOutgoingSieveScriptBefore() throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, "");
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * outgoing sieve script defined by admin (not able to edit and view from
+     * the end user) applied before the end user filter rule
+     *
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2092)
+    public Map<String,Object> unsetMailAdminOutgoingSieveScriptBefore(Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminOutgoingSieveScriptBefore, "");
+        return attrs;
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied after the end user filter rule
+     *
+     * @return zimbraMailAdminSieveScriptAfter, or null if unset
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2091)
+    public String getMailAdminSieveScriptAfter() {
+        return getAttr(Provisioning.A_zimbraMailAdminSieveScriptAfter, null, true);
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied after the end user filter rule
+     *
+     * @param zimbraMailAdminSieveScriptAfter new value
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2091)
+    public void setMailAdminSieveScriptAfter(String zimbraMailAdminSieveScriptAfter) throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, zimbraMailAdminSieveScriptAfter);
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied after the end user filter rule
+     *
+     * @param zimbraMailAdminSieveScriptAfter new value
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2091)
+    public Map<String,Object> setMailAdminSieveScriptAfter(String zimbraMailAdminSieveScriptAfter, Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, zimbraMailAdminSieveScriptAfter);
+        return attrs;
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied after the end user filter rule
+     *
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2091)
+    public void unsetMailAdminSieveScriptAfter() throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, "");
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied after the end user filter rule
+     *
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2091)
+    public Map<String,Object> unsetMailAdminSieveScriptAfter(Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptAfter, "");
+        return attrs;
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied before the end user filter rule
+     *
+     * @return zimbraMailAdminSieveScriptBefore, or null if unset
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2090)
+    public String getMailAdminSieveScriptBefore() {
+        return getAttr(Provisioning.A_zimbraMailAdminSieveScriptBefore, null, true);
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied before the end user filter rule
+     *
+     * @param zimbraMailAdminSieveScriptBefore new value
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2090)
+    public void setMailAdminSieveScriptBefore(String zimbraMailAdminSieveScriptBefore) throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, zimbraMailAdminSieveScriptBefore);
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied before the end user filter rule
+     *
+     * @param zimbraMailAdminSieveScriptBefore new value
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2090)
+    public Map<String,Object> setMailAdminSieveScriptBefore(String zimbraMailAdminSieveScriptBefore, Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, zimbraMailAdminSieveScriptBefore);
+        return attrs;
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied before the end user filter rule
+     *
+     * @throws com.zimbra.common.service.ServiceException if error during update
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2090)
+    public void unsetMailAdminSieveScriptBefore() throws com.zimbra.common.service.ServiceException {
+        HashMap<String,Object> attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, "");
+        getProvisioning().modifyAttrs(this, attrs);
+    }
+
+    /**
+     * sieve script defined by admin (not able to edit and view from the end
+     * user) applied before the end user filter rule
+     *
+     * @param attrs existing map to populate, or null to create a new map
+     * @return populated map to pass into Provisioning.modifyAttrs
+     *
+     * @since ZCS 8.7.7
+     */
+    @ZAttr(id=2090)
+    public Map<String,Object> unsetMailAdminSieveScriptBefore(Map<String,Object> attrs) {
+        if (attrs == null) attrs = new HashMap<String,Object>();
+        attrs.put(Provisioning.A_zimbraMailAdminSieveScriptBefore, "");
+        return attrs;
+    }
+
+    /**
      * Whether to allow password sent to non-secured port when zimbraMailMode
      * is mixed. If it set to TRUE the server will allow login with clear
      * text AuthRequests and change password with clear text
@@ -28232,7 +28463,7 @@ public abstract class ZAttrConfig extends Entry {
      *
      * @return zimbraMailSieveNotifyActionRFCCompliant, or false if unset
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2095)
     public boolean isMailSieveNotifyActionRFCCompliant() {
@@ -28248,7 +28479,7 @@ public abstract class ZAttrConfig extends Entry {
      * @param zimbraMailSieveNotifyActionRFCCompliant new value
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2095)
     public void setMailSieveNotifyActionRFCCompliant(boolean zimbraMailSieveNotifyActionRFCCompliant) throws com.zimbra.common.service.ServiceException {
@@ -28267,7 +28498,7 @@ public abstract class ZAttrConfig extends Entry {
      * @param attrs existing map to populate, or null to create a new map
      * @return populated map to pass into Provisioning.modifyAttrs
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2095)
     public Map<String,Object> setMailSieveNotifyActionRFCCompliant(boolean zimbraMailSieveNotifyActionRFCCompliant, Map<String,Object> attrs) {
@@ -28284,7 +28515,7 @@ public abstract class ZAttrConfig extends Entry {
      *
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2095)
     public void unsetMailSieveNotifyActionRFCCompliant() throws com.zimbra.common.service.ServiceException {
@@ -28302,7 +28533,7 @@ public abstract class ZAttrConfig extends Entry {
      * @param attrs existing map to populate, or null to create a new map
      * @return populated map to pass into Provisioning.modifyAttrs
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2095)
     public Map<String,Object> unsetMailSieveNotifyActionRFCCompliant(Map<String,Object> attrs) {
@@ -65363,89 +65594,12 @@ public abstract class ZAttrConfig extends Entry {
     }
 
     /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the user-defined sieve rule.
-     *
-     * @return zimbraSieveFeatureVariablesEnabled, or false if unset
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2096)
-    public boolean isSieveFeatureVariablesEnabled() {
-        return getBooleanAttr(Provisioning.A_zimbraSieveFeatureVariablesEnabled, false, true);
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the user-defined sieve rule.
-     *
-     * @param zimbraSieveFeatureVariablesEnabled new value
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2096)
-    public void setSieveFeatureVariablesEnabled(boolean zimbraSieveFeatureVariablesEnabled) throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraSieveFeatureVariablesEnabled, zimbraSieveFeatureVariablesEnabled ? Provisioning.TRUE : Provisioning.FALSE);
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the user-defined sieve rule.
-     *
-     * @param zimbraSieveFeatureVariablesEnabled new value
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2096)
-    public Map<String,Object> setSieveFeatureVariablesEnabled(boolean zimbraSieveFeatureVariablesEnabled, Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraSieveFeatureVariablesEnabled, zimbraSieveFeatureVariablesEnabled ? Provisioning.TRUE : Provisioning.FALSE);
-        return attrs;
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the user-defined sieve rule.
-     *
-     * @throws com.zimbra.common.service.ServiceException if error during update
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2096)
-    public void unsetSieveFeatureVariablesEnabled() throws com.zimbra.common.service.ServiceException {
-        HashMap<String,Object> attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraSieveFeatureVariablesEnabled, "");
-        getProvisioning().modifyAttrs(this, attrs);
-    }
-
-    /**
-     * Whether to enable the Sieve &quot;Variables&quot; extension defined in
-     * RFC 5229 in the user-defined sieve rule.
-     *
-     * @param attrs existing map to populate, or null to create a new map
-     * @return populated map to pass into Provisioning.modifyAttrs
-     *
-     * @since ZCS 8.7.6
-     */
-    @ZAttr(id=2096)
-    public Map<String,Object> unsetSieveFeatureVariablesEnabled(Map<String,Object> attrs) {
-        if (attrs == null) attrs = new HashMap<String,Object>();
-        attrs.put(Provisioning.A_zimbraSieveFeatureVariablesEnabled, "");
-        return attrs;
-    }
-
-    /**
      * Whether to enable the Sieve &quot;reject&quot; action defined in RFC
      * 5429.
      *
      * @return zimbraSieveRejectEnabled, or true if unset
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2094)
     public boolean isSieveRejectEnabled() {
@@ -65459,7 +65613,7 @@ public abstract class ZAttrConfig extends Entry {
      * @param zimbraSieveRejectEnabled new value
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2094)
     public void setSieveRejectEnabled(boolean zimbraSieveRejectEnabled) throws com.zimbra.common.service.ServiceException {
@@ -65476,7 +65630,7 @@ public abstract class ZAttrConfig extends Entry {
      * @param attrs existing map to populate, or null to create a new map
      * @return populated map to pass into Provisioning.modifyAttrs
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2094)
     public Map<String,Object> setSieveRejectEnabled(boolean zimbraSieveRejectEnabled, Map<String,Object> attrs) {
@@ -65491,7 +65645,7 @@ public abstract class ZAttrConfig extends Entry {
      *
      * @throws com.zimbra.common.service.ServiceException if error during update
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2094)
     public void unsetSieveRejectEnabled() throws com.zimbra.common.service.ServiceException {
@@ -65507,7 +65661,7 @@ public abstract class ZAttrConfig extends Entry {
      * @param attrs existing map to populate, or null to create a new map
      * @return populated map to pass into Provisioning.modifyAttrs
      *
-     * @since ZCS 8.7.6
+     * @since ZCS 8.7.7
      */
     @ZAttr(id=2094)
     public Map<String,Object> unsetSieveRejectEnabled(Map<String,Object> attrs) {
