+++ store/src/java/com/zimbra/cs/filter/jsieve/Notify.java
@@ -28,6 +28,9 @@ import org.apache.jsieve.exception.SieveException;
 import org.apache.jsieve.exception.SyntaxException;
 import org.apache.jsieve.mail.MailAdapter;
 
+import com.zimbra.common.util.ZimbraLog;
+import com.zimbra.cs.filter.JsieveConfigMapHandler;
+
 import java.util.List;
 
 public class Notify extends AbstractActionCommand {
